+++ packages/core/src/config/options.ts
