+++ packages/core/src/database/cache/sqliteCacheStore.ts
@@ -1,7 +1,7 @@
 import type Sqlite from "better-sqlite3";
 
 import { merge_intervals } from "@/common/utils";
-import type { Block, Log, Transaction } from "@/types";
+import type { Block, Log, Transaction } from "@/database/types";
 
 import type { CachedInterval, CacheStore, ContractCall } from "./cacheStore";
