+++ packages/core/src/database/entity/utils.ts
