+++ packages/core/src/index.ts
@@ -1,9 +1,6 @@
 export type { ReadOnlyContract } from "@/codegen/ethers-abitype";
-export type { PonderOptions } from "@/common/options";
-export type {
-  PonderConfig,
-  ResolvedPonderConfig,
-} from "@/config/buildPonderConfig";
+export type { PonderOptions } from "@/config/options";
+export type { PonderConfig, ResolvedPonderConfig } from "@/config/ponderConfig";
+export type { Block, Log, Transaction } from "@/database/types";
 export { Ponder } from "@/Ponder";
 export { PonderApp } from "@/reload/readHandlers";
-export type { Block, Log, Transaction } from "@/types";
